==> Utilities/10-Normalisation.R <==
#!/usr/bin/env Rscript
#SBATCH -J Normalisation
#SBATCH -o Normalisation.%j.out
#SBATCH -e Normalisation.%j.err
#SBATCH --mincpus 8 
#SBATCH --mem=16G
#SBATCH --time=02:57:42

# This script normalizes the full data set and the 500 cells/sample
# data set generated by the script 09-QC_and_Filtering.R
library(scater)
library(scran)
library(tidyverse)
library(BiocSingular)
library(BiocParallel)

bpp <- MulticoreParam(8)

## Full data set

# Load the data
sce <- readRDS("data/R_objects/Caron_filtered.full.rds")

# cluster cells
set.seed(100) 
clust <- quickCluster(sce) 

# calculate size factors 
sce <- computePooledFactors(sce, cluster=clust, min.mean=0.1)

# log Normalize the Counts
sce <- logNormCounts(sce)

# save normalised object
saveRDS(sce, "data/R_objects/Caron_normalized.full.rds")

## 500 cell/sample data set

# Load the data
sce <- readRDS("data/R_objects/Caron_filtered.500.rds")

# cluster cells
set.seed(100) 
clust <- quickCluster(sce) 

# calculate size factors 
sce <- computePooledFactors(sce, cluster=clust, min.mean=0.1)

# log Normalize the Counts
sce <- logNormCounts(sce)

# save normalised object
saveRDS(sce, "data/R_objects/Caron_normalized.500.rds")
